requirements.txt
httplib2
pytz
python-telegram-bot
schedule
google-api-python-client
oauth2client
sentry-sdk==0.6.8
